Join and merge files of `FreeC` and `Algebra`.
As part of the work towards the simplifed model of `fs3`, we unify the
files and objects `FreeC` and `Algebra` into a single one.
This allows us to re-seal the `FreeC` GADT, since it is in one file.

To reduce some collisions, we also do these smaller changes:

- For convenience of a smaller diff and preserving source history, we
   move lines from `FreeC.scala` to `Algebra.scala`, as there are fewer.
- Due to collision between `FreeC.Eval`, and `Algebra.Eval`, we rename
   the latter into `Lift`, in the sense of "lifting"
   an F<A> into a Stream<F, A>.
- To reduce changes in the `Stream` and `Pull` files, we keep
  `Algebra` as an internal alias of `FreeC`.
