+++ core/shared/src/main/scala-2.13/internal/package.scala
@@ -1,5 +1,6 @@
 package fs2
 
 package object internal {
+  private[fs2] val Algebra = FreeC
   private[fs2] type Factory[-A, +C] = scala.collection.Factory[A, C]
 }
